Update JigsawEvent to 1.18
