+++ patches/minecraft/net/minecraft/world/level/levelgen/feature/structures/JigsawPlacement.java.patch
@@ -1,0 +1,73 @@
+--- a/net/minecraft/world/level/levelgen/feature/structures/JigsawPlacement.java
++++ b/net/minecraft/world/level/levelgen/feature/structures/JigsawPlacement.java
+@@ -43,6 +_,10 @@
+    static final Logger f_68941_ = LogManager.getLogger();
+ 
+    public static Optional<PieceGenerator<JigsawConfiguration>> m_197210_(PieceGeneratorSupplier.Context<JigsawConfiguration> p_197211_, JigsawPlacement.PieceFactory p_197212_, BlockPos p_197213_, boolean p_197214_, boolean p_197215_) {
++      return addPieces(p_197211_, com.google.common.collect.Maps.newHashMap(), p_197212_, p_197213_, p_197214_, p_197215_);
++   }
++
++   public static Optional<PieceGenerator<JigsawConfiguration>> addPieces(PieceGeneratorSupplier.Context<JigsawConfiguration> p_197211_, java.util.Map<ResourceLocation, Object> placerSettings, JigsawPlacement.PieceFactory p_197212_, BlockPos p_197213_, boolean p_197214_, boolean p_197215_) {
+       WorldgenRandom worldgenrandom = new WorldgenRandom(new LegacyRandomSource(0L));
+       worldgenrandom.m_190068_(p_197211_.f_197354_(), p_197211_.f_197355_().f_45578_, p_197211_.f_197355_().f_45579_);
+       RegistryAccess registryaccess = p_197211_.f_197360_();
+@@ -81,7 +_,7 @@
+                if (jigsawconfiguration.m_67765_() > 0) {
+                   int i1 = 80;
+                   AABB aabb = new AABB((double)(i - 80), (double)(k - 80), (double)(j - 80), (double)(i + 80 + 1), (double)(k + 80 + 1), (double)(j + 80 + 1));
+-                  JigsawPlacement.Placer jigsawplacement$placer = new JigsawPlacement.Placer(registry, jigsawconfiguration.m_67765_(), p_197212_, chunkgenerator, structuremanager, list, worldgenrandom);
++                  JigsawPlacement.Placer jigsawplacement$placer = new JigsawPlacement.Placer(registry, jigsawconfiguration.m_67765_(), placerSettings, p_197212_, chunkgenerator, structuremanager, list, worldgenrandom);
+                   jigsawplacement$placer.f_69001_.addLast(new JigsawPlacement.PieceState(poolelementstructurepiece, new MutableObject<>(Shapes.m_83113_(Shapes.m_83064_(aabb), Shapes.m_83064_(AABB.m_82321_(boundingbox)), BooleanOp.f_82685_)), 0));
+ 
+                   while(!jigsawplacement$placer.f_69001_.isEmpty()) {
+@@ -97,8 +_,12 @@
+    }
+ 
+    public static void m_161624_(RegistryAccess p_161625_, PoolElementStructurePiece p_161626_, int p_161627_, JigsawPlacement.PieceFactory p_161628_, ChunkGenerator p_161629_, StructureManager p_161630_, List<? super PoolElementStructurePiece> p_161631_, Random p_161632_, LevelHeightAccessor p_161633_) {
++      addPieces(p_161625_, p_161626_, p_161627_, com.google.common.collect.Maps.newHashMap(), p_161628_, p_161629_, p_161630_, p_161631_, p_161632_, p_161633_);
++   }
++
++   public static void addPieces(RegistryAccess p_161625_, PoolElementStructurePiece p_161626_, int p_161627_, java.util.Map<ResourceLocation, Object> placerSettings, JigsawPlacement.PieceFactory p_161628_, ChunkGenerator p_161629_, StructureManager p_161630_, List<? super PoolElementStructurePiece> p_161631_, Random p_161632_, LevelHeightAccessor p_161633_) {
+       Registry<StructureTemplatePool> registry = p_161625_.m_175515_(Registry.f_122884_);
+-      JigsawPlacement.Placer jigsawplacement$placer = new JigsawPlacement.Placer(registry, p_161627_, p_161628_, p_161629_, p_161630_, p_161631_, p_161632_);
++      JigsawPlacement.Placer jigsawplacement$placer = new JigsawPlacement.Placer(registry, p_161627_, placerSettings, p_161628_, p_161629_, p_161630_, p_161631_, p_161632_);
+       jigsawplacement$placer.f_69001_.addLast(new JigsawPlacement.PieceState(p_161626_, new MutableObject<>(Shapes.f_83036_), 0));
+ 
+       while(!jigsawplacement$placer.f_69001_.isEmpty()) {
+@@ -127,6 +_,7 @@
+    static final class Placer {
+       private final Registry<StructureTemplatePool> f_68994_;
+       private final int f_68995_;
++      private final java.util.Map<ResourceLocation, Object> placerSettings;
+       private final JigsawPlacement.PieceFactory f_68996_;
+       private final ChunkGenerator f_68997_;
+       private final StructureManager f_68998_;
+@@ -134,9 +_,10 @@
+       private final Random f_69000_;
+       final Deque<JigsawPlacement.PieceState> f_69001_ = Queues.newArrayDeque();
+ 
+-      Placer(Registry<StructureTemplatePool> p_69003_, int p_69004_, JigsawPlacement.PieceFactory p_69005_, ChunkGenerator p_69006_, StructureManager p_69007_, List<? super PoolElementStructurePiece> p_69008_, Random p_69009_) {
++      Placer(Registry<StructureTemplatePool> p_69003_, int p_69004_, java.util.Map<ResourceLocation, Object> placerSettings, JigsawPlacement.PieceFactory p_69005_, ChunkGenerator p_69006_, StructureManager p_69007_, List<? super PoolElementStructurePiece> p_69008_, Random p_69009_) {
+          this.f_68994_ = p_69003_;
+          this.f_68995_ = p_69004_;
++         this.placerSettings = placerSettings;
+          this.f_68996_ = p_69005_;
+          this.f_68997_ = p_69006_;
+          this.f_68998_ = p_69007_;
+@@ -181,6 +_,7 @@
+                   List<StructurePoolElement> list = Lists.newArrayList();
+                   if (p_191515_ != this.f_68995_) {
+                      list.addAll(optional.get().m_69276_(this.f_69000_));
++                     net.minecraftforge.common.MinecraftForge.EVENT_BUS.post(new net.minecraftforge.event.world.JigsawPieceEvent(list, resourcelocation, blockpos1, this.placerSettings, this.f_68994_, this.f_69000_, p_191515_));
+                   }
+ 
+                   list.addAll(optional1.get().m_69276_(this.f_69000_));
+@@ -246,7 +_,7 @@
+                                  boundingbox3.m_162371_(new BlockPos(boundingbox3.m_162395_(), boundingbox3.m_162396_() + j2, boundingbox3.m_162398_()));
+                               }
+ 
+-                              if (!Shapes.m_83157_(mutableobject1.getValue(), Shapes.m_83064_(AABB.m_82321_(boundingbox3).m_82406_(0.25D)), BooleanOp.f_82683_)) {
++                              if ((boolean) placerSettings.getOrDefault(new ResourceLocation("forge:force_piece_placement"), false) || !Shapes.m_83157_(mutableobject1.getValue(), Shapes.m_83064_(AABB.m_82321_(boundingbox3).m_82406_(0.25D)), BooleanOp.f_82683_)) {
+                                  mutableobject1.setValue(Shapes.m_83148_(mutableobject1.getValue(), Shapes.m_83064_(AABB.m_82321_(boundingbox3)), BooleanOp.f_82685_));
+                                  int i3 = p_191513_.m_72647_();
+                                  int k2;
