+++ src/main/java/net/minecraftforge/event/level/JigsawPieceEvent.java
@@ -1,10 +1,10 @@
 package net.minecraftforge.event.world;
 
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.registry.Registry;
-import net.minecraft.world.gen.feature.jigsaw.JigsawPattern;
-import net.minecraft.world.gen.feature.jigsaw.JigsawPiece;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Registry;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.levelgen.feature.structures.StructurePoolElement;
+import net.minecraft.world.level.levelgen.feature.structures.StructureTemplatePool;
 import net.minecraftforge.eventbus.api.Cancelable;
 import net.minecraftforge.eventbus.api.Event;
 
@@ -13,9 +13,9 @@ import java.util.Map;
 import java.util.Random;
 
 /**
- * Is fired from {@link net.minecraft.world.gen.feature.jigsaw.JigsawManager.Assembler#tryPlacingChildren}
+ * Is fired from {@link net.minecraft.world.level.levelgen.feature.structures.JigsawPlacement.Placer#tryPlacingChildren}
  * for each generated Jigsaw block after it stores the pool of Jigsaw pieces in can attach to.
- * The list does NOT contain the pieces contained in {@link net.minecraft.world.gen.feature.jigsaw.JigsawPattern#fallback}
+ * The list does NOT contain the pieces contained in {@link net.minecraft.world.level.levelgen.feature.structures.StructureTemplatePool#fallback}
  * to avoid infinite generation problems.
  * <br><br>
  * Adding a custom villager's house to plains village and zombie village can be just:
@@ -23,7 +23,7 @@ import java.util.Random;
  * {@code
  *    String pool = event.patternLocation.getPath();
  *    if (pool.startsWith("village/plains/houses") || pool.startsWith("village/plains/zombie/houses"))
- *      event.pieces.add(new SingleJigsawPiece("namespace:path"));
+ *      event.pieces.add(StructurePoolElement.single("namespace:path"));
  * }
  *
  * For more detailed insight see <a href=https://mcforge.readthedocs.io/en/1.16.x/events/jigsawevent/>here.</a>
@@ -31,33 +31,33 @@ import java.util.Random;
 @Cancelable
 public class JigsawPieceEvent extends Event
 {
-   public final List<JigsawPiece> pieces;
+   public final List<StructurePoolElement> pieces;
    public final ResourceLocation patternLocation;
    public final BlockPos jigsawBlockPosition;
-   public final Map<ResourceLocation, Object> jigsawAssemblerSettings;
-   public final Registry<JigsawPattern> patternRegistry;
+   public final Map<ResourceLocation, Object> jigsawPlacerSettings;
+   public final Registry<StructureTemplatePool> patternRegistry;
    public final Random random;
    public final int currentDepth;
 
    /**
     * @param pieces                  The list of pieces that have a probability of being placed. Any addition to big pools should be shuffled for better results.
-    *                                The list does NOT contain the fallback pool of the current pool. See the appropriate pool for the structure to know what the fallback.
-    *                                ({@link net.minecraft.world.gen.feature.structure.PlainsVillagePools} for Plains Village)
-    * @param patternLocation         The resource location of the {@link net.minecraft.world.gen.feature.jigsaw.JigsawPattern}.
+    *                                The list does NOT contain the fallback pool of the current pool. See the appropriate pool for the structure to know what the fallback is.
+    *                                ({@link net.minecraft.data.worldgen.PlainVillagePools} for Plains Village)
+    * @param patternLocation         The resource location of the {@link net.minecraft.world.level.levelgen.feature.structures.StructureTemplatePool}.
     * @param jigsawBlockPosition     The block position of jigsaw block requesting the pieces.
-    * @param jigsawAssemblerSettings The map of generate-time settings for specific structure. It's unused for vanilla structures but can be used
+    * @param jigsawPlacerSettings    The map of generate-time settings for specific structure. It's unused for vanilla structures but can be used
     *                                in modded ones to add specific settings or more or less randomness. If `forge:force_piece_placement` is set to `true`,
     *                                all future pieces in that structure will always be placed, even if they overwrite existing pieces.
     * @param patternRegistry         The access to Registry instance for structure pools. Can be used to get pools of other structures and re-use them.
     * @param random                  The access to Random instance of specific structure. Can be used to shuffle custom pieces after they are added.
     * @param currentDepth            The depth of the piece that is going to be placed. It will be 0 for the jigsaw blocks contained in the starting piece.
     */
-   public JigsawPieceEvent(List<JigsawPiece> pieces, ResourceLocation patternLocation, BlockPos jigsawBlockPosition, Map<ResourceLocation, Object> jigsawAssemblerSettings, Registry<JigsawPattern> patternRegistry, Random random, int currentDepth)
+   public JigsawPieceEvent(List<StructurePoolElement> pieces, ResourceLocation patternLocation, BlockPos jigsawBlockPosition, Map<ResourceLocation, Object> jigsawPlacerSettings, Registry<StructureTemplatePool> patternRegistry, Random random, int currentDepth)
    {
       this.pieces = pieces;
       this.patternLocation = patternLocation;
       this.jigsawBlockPosition = jigsawBlockPosition;
-      this.jigsawAssemblerSettings = jigsawAssemblerSettings;
+      this.jigsawPlacerSettings = jigsawPlacerSettings;
       this.patternRegistry = patternRegistry;
       this.random = random;
       this.currentDepth = currentDepth;
